Add files via upload
This is the notebook that I created to produce the Pandemic in 60 seconds video that I posted to Reddit.  The notebook was used on my laptop - a WINDOWS laptop and it uses the OS package to generate a list of files - so if you are on a different platform, you might have to make changes.  Also, note that the the jpeg creation requires Orca to be installed.  Just search for it and use conda to install it from the conda console prompt.  Other methods are a pain.  You'll need the Plotly Express, pandas and CV2 packages.

This one downloads from the NYT data.  It also grabs a Census file that I created and put in Git.