Add files via upload
Updated 5/31/2022 due to changes in the source data provided by the New York Times.  Prior to mid-May 2022, they used a single dataset for the county-level data.  After, they divided the data into 2020,2021, and 2022 datasets.  This code processes the data.

Note that it is set up with a code block deactivated.  That code block is used to download the 2020 and 2021 data and make a local copy of the data.  As this data does not change, downloading it each time is a waste of time.  Activate the block one time for the first run, then deactivate it.  Note that there are physical locations hardcoded that you will need to adjust to your own environment.